refactor(#157): 친구 요청 훅 추출
