+++ src/features/user/buttons/friend-request-send-button.tsx
@@ -1,8 +1,4 @@
-import { useMutation } from "@tanstack/react-query";
-import { userKeys } from "@/entities/user/config/query-keys";
-import { api, type OtherProfileResponse } from "@/shared/api";
-import { queryClient } from "@/shared/lib/query-client";
-import { toast } from "@/shared/lib/toast";
+import { useSendFriendRequest } from "@/features/user/hooks/use-send-friend-request";
 import { useAuth } from "@/shared/model/use-auth";
 import { Button } from "@/shared/ui";
 
@@ -12,37 +8,7 @@ export default function FriendRequestSendButton({
 	userId: number;
 }) {
 	const { user } = useAuth();
-
-	const addFriendMutation = useMutation({
-		mutationFn: () => api.private.friend.sendFriendRequest(userId),
-		onMutate: async () => {
-			await queryClient.cancelQueries({
-				queryKey: userKeys.profile(userId),
-			});
-			const previous = queryClient.getQueryData(userKeys.profile(userId));
-
-			queryClient.setQueryData(
-				userKeys.profile(userId),
-				(old: OtherProfileResponse) => ({
-					...old,
-					friendRequestMemberId: user?.id,
-				}),
-			);
-
-			return { previous };
-		},
-		onError: (err, _, context) => {
-			console.log(err);
-			queryClient.setQueryData(userKeys.profile(userId), context?.previous);
-			toast.error("친구 요청이 실패했습니다.");
-		},
-		onSuccess: () => {
-			toast.confirm("친구 요청을 보냈습니다.");
-		},
-		onSettled: () => {
-			queryClient.invalidateQueries({ queryKey: userKeys.profile(userId) });
-		},
-	});
+	const { mutate } = useSendFriendRequest(userId);
 
 	if (!user?.id) return null;
 
@@ -51,7 +17,7 @@ export default function FriendRequestSendButton({
 			variant={"black"}
 			size="xl"
 			className="bold-14 h-[45px] mobile:w-[218px] w-full mobile:rounded-xl rounded-[6px]"
-			onClick={() => addFriendMutation.mutate()}
+			onClick={() => mutate()}
 		>
 			친구 추가
 		</Button>
